finsihed HW6 both coding qestions React Fundamentals
